+++ src/test/java/org/apache/commons/io/input/ReversedLinesFileReaderTestParamFile.java
@@ -37,28 +37,27 @@ import org.junit.runners.Parameterized.Parameters;
 
 /**
  * Test checks symmetric behaviour with  BufferedReader
- *
  */
 @RunWith(Parameterized.class)
 public class ReversedLinesFileReaderTestParamFile {
 
-    @Parameters(name = "{0}, charset={1}") 
+    @Parameters(name = "{0}, charset={1}")
     public static Collection<Object[]> blockSizes() {
-            return Arrays.asList(new Object[][] {
-                    {"test-file-20byteslength.bin", "ISO_8859_1", null},
-                    {"test-file-iso8859-1-shortlines-win-linebr.bin", "ISO_8859_1", null},
-                    {"test-file-iso8859-1.bin", "ISO_8859_1", null},
-                    {"test-file-shiftjis.bin", "Shift_JIS", null},
-                    {"test-file-utf16be.bin", "UTF-16BE", null},
-                    {"test-file-utf16le.bin", "UTF-16LE", null},
-                    {"test-file-utf8-cr-only.bin", "UTF-8", null},
-                    {"test-file-utf8-win-linebr.bin", "UTF-8", null},
-                    {"test-file-utf8-win-linebr.bin", "UTF-8", Integer.valueOf(1)},
-                    {"test-file-utf8-win-linebr.bin", "UTF-8", Integer.valueOf(2)},
-                    {"test-file-utf8-win-linebr.bin", "UTF-8", Integer.valueOf(3)},
-                    {"test-file-utf8-win-linebr.bin", "UTF-8", Integer.valueOf(4)},
-                    {"test-file-utf8.bin", "UTF-8", null},
-            });
+        return Arrays.asList(new Object[][]{
+                {"test-file-20byteslength.bin", "ISO_8859_1", null},
+                {"test-file-iso8859-1-shortlines-win-linebr.bin", "ISO_8859_1", null},
+                {"test-file-iso8859-1.bin", "ISO_8859_1", null},
+                {"test-file-shiftjis.bin", "Shift_JIS", null},
+                {"test-file-utf16be.bin", "UTF-16BE", null},
+                {"test-file-utf16le.bin", "UTF-16LE", null},
+                {"test-file-utf8-cr-only.bin", "UTF-8", null},
+                {"test-file-utf8-win-linebr.bin", "UTF-8", null},
+                {"test-file-utf8-win-linebr.bin", "UTF-8", 1},
+                {"test-file-utf8-win-linebr.bin", "UTF-8", 2},
+                {"test-file-utf8-win-linebr.bin", "UTF-8", 3},
+                {"test-file-utf8-win-linebr.bin", "UTF-8", 4},
+                {"test-file-utf8.bin", "UTF-8", null},
+        });
     }
 
     private ReversedLinesFileReader reversedLinesFileReader;
@@ -71,12 +70,12 @@ public class ReversedLinesFileReaderTestParamFile {
     public ReversedLinesFileReaderTestParamFile(final String fileName, final String encoding, final Integer buffsize) {
         this.fileName = fileName;
         this.encoding = encoding;
-        this.buffSize = buffsize == null ? 4096 : buffsize.intValue();
+        this.buffSize = buffsize == null ? 4096 : buffsize;
     }
 
     @Test
     public void testDataIntegrityWithBufferedReader() throws URISyntaxException, IOException {
-        final File testFileIso = new File(this.getClass().getResource("/"+fileName).toURI());
+        final File testFileIso = new File(this.getClass().getResource("/" + fileName).toURI());
         reversedLinesFileReader = new ReversedLinesFileReader(testFileIso, buffSize, encoding);
 
         final Stack<String> lineStack = new Stack<String>();
@@ -85,12 +84,12 @@ public class ReversedLinesFileReaderTestParamFile {
         String line = null;
 
         // read all lines in normal order
-        while((line = bufferedReader.readLine())!=null) {
+        while ((line = bufferedReader.readLine()) != null) {
             lineStack.push(line);
         }
 
         // read in reverse order and compare with lines from stack
-        while((line = reversedLinesFileReader.readLine())!=null) {
+        while ((line = reversedLinesFileReader.readLine()) != null) {
             final String lineFromBufferedReader = lineStack.pop();
             assertEquals(lineFromBufferedReader, line);
         }
@@ -101,16 +100,15 @@ public class ReversedLinesFileReaderTestParamFile {
     public void closeReader() {
         try {
             bufferedReader.close();
-        } catch(final Exception e) {
+        } catch (final Exception e) {
             // ignore
         }
         try {
             reversedLinesFileReader.close();
-        } catch(final Exception e) {
+        } catch (final Exception e) {
             // ignore
         }
     }
-
 
 
 }
